requirements.txt
faiss-cpu
numpy
torch
transformers
sentence-transformers
nltk
scipy
streamlit
pypdf2